Dar formato a los valores numericos

+++ app/src/main/java/com/example/proyectointermodular/modelo/FacturaRecibida.kt
@@ -13,7 +13,7 @@ data class FacturaRecibida(
 
     // Datos financieros
     val baseImponible: Double = 0.00,
-    val tipoIva: Double = 0.00,
+    val tipoIva: Int = 0,
     val cuotaIva: Double = 0.00,
     val total: Double = 0.00,
